Restructured methods
