+++ GoServer/controllers/listings/get_listing_by_item_id.go
@@ -1,0 +1,42 @@
+package listings
+
+import (
+	"net/http"
+
+	"github.com/aaronangxz/TIC2601/models"
+	"github.com/aaronangxz/TIC2601/utils"
+
+	"github.com/gin-gonic/gin"
+)
+
+func GetListingByItemID(c *gin.Context) {
+	var (
+		singleListing models.Listing
+		input         models.GetSingleListingRequest
+	)
+
+	if err := c.ShouldBindJSON(&input); err != nil {
+		if input.ItemID == nil {
+			c.JSON(http.StatusBadRequest, gin.H{"Respmeta": models.NewParamErrorsResponse("item_id cannot be empty.")})
+			return
+		}
+		if !utils.ValidateUint(input.ItemID) {
+			c.JSON(http.StatusBadRequest, gin.H{"Respmeta": models.NewParamErrorsResponse("item_id must be uint type.")})
+			return
+		}
+		c.JSON(http.StatusBadRequest, gin.H{"Respmeta": models.NewJSONErrorResponse(err)})
+		return
+	}
+
+	if input.GetItemID() == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{"Respmeta": models.NewParamErrorsResponse("item_id must be > 0.")})
+		return
+	}
+
+	if err := models.DB.Where("item_id = ?", input.ItemID).First(&singleListing).Error; err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"Respmeta": models.NewDBErrorResponse(err)})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"Data": singleListing})
+}
